backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
sqlite-utils==3.35.2

# Document Processing
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
python-magic==0.4.27
pypdf==3.17.0

# OCR and Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.2

# LLM Integration (using Anthropic Claude for lightweight/cost-effective processing)
anthropic==0.7.8
openai==1.3.7
tiktoken==0.5.2

# Content Safety & PII Detection
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
python-json-logger==2.0.7
pyyaml==6.0.1
tqdm==4.66.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
